ch2: lilypond-custom: music entry: initial PDF
This is the _very plain_ music entry with LilyPond
2.19.1 latest master.

I think that particularly the piano centered dynamics made
it quite ugly.

Next step is to check against the Clara Schumann score,
as there are plenty of differences.

Then I'll have to fix directions of markup, a slur and several
articulations.

Then check the reminder accidentals.